Make cleanup conditionally

+++ runtime/src/main/java/de/prgrm/topology/runtime/service/ServiceRegistry.java
@@ -61,7 +61,7 @@ public class ServiceRegistry {
     }
 
     void onStop(@Observes ShutdownEvent event) {
-        if (config.enabled()) {
+        if (config.enabled() && config.cleanupOnShutdown()) {
             deleteLocalTopology();
         }
     }
